code structure change to make enterprise more standalone and reduce duplication
